fix(settlement): rename api.rs files to be explicit about which API they refer to
