+++ crates/interledger-settlement/src/settlement_api/mod.rs
@@ -1,6 +1,6 @@
-mod api;
 mod client;
 mod message_service;
+mod node_api;
 
 #[cfg(test)]
 mod fixtures;
@@ -8,6 +8,6 @@ mod fixtures;
 mod test_helpers;
 
 // Expose the API creation filter method and the necessary services
-pub use api::create_settlements_filter;
 pub use client::SettlementClient;
 pub use message_service::SettlementMessageService;
+pub use node_api::create_settlements_filter;
